Move remaining components to folders and update paths
